refactor(scripts): rename deps to constants, SCREAMING_CASE
