+++ scripts/generate/labels.ts
@@ -1,6 +1,6 @@
 import * as path from "@std/path";
 
-import { REPO_ROOT } from "@/deps.ts";
+import { REPO_ROOT } from "../constants.ts";
 import { updateFile } from "@/generate/utils.ts";
 import { UserStylesSchema } from "@/types/mod.ts";
 import * as yaml from "@std/yaml";
