+++ scripts/lint/file-checker.ts
@@ -2,7 +2,7 @@ import * as fs from "@std/fs";
 import * as path from "@std/path";
 import core from "@actions/core";
 
-import { REPO_ROOT } from "@/deps.ts";
+import { REPO_ROOT } from "../constants.ts";
 import { log } from "@/logger.ts";
 
 const requiredFiles = [
